+++ frontend/src/components/Landing/Footer.js
