+++ frontend/src/components/Landing/Image.js
