+++ frontend/src/pages/index.js
@@ -1,8 +1,8 @@
 import React from "react"
 
-import Layout from "../components/Landing2/Layout2"
+import Layout from "../components/Landing/Layout"
 import SEO from "../components/seo"
-import Landing from "../components/Landing2/Landing"
+import Landing from "../components/Landing/Landing"
 
 const IndexPage = () => (
   <Layout>
